refactor namespaces for internal encapsulation
Reorganized namespaces to move classes and interfaces from
`Nalix.Network.Routing.Results` and `Nalix.Network.Routing.Metadata`
to `Nalix.Network.Internal.Results` and `Nalix.Network.Internal.Routing`.

Updated all relevant files to reflect the new namespace structure:
- Replaced `Nalix.Network.Routing.Results.*` with `Nalix.Network.Internal.Results.*`.
- Replaced `Nalix.Network.Routing.Metadata` with `Nalix.Network.Internal.Routing`.

This restructuring encapsulates packet handling, routing, and result-related
components as internal implementation details, improving modularity and
enforcing better separation of concerns within the codebase.
