+++ src/Nalix.Network/Internal/Results/Memory/ReadOnlyMemoryReturnHandler.cs
@@ -6,8 +6,9 @@ using System.ComponentModel;
 using System.Threading.Tasks;
 using Nalix.Common.Networking.Packets;
 using Nalix.Common.Networking.Protocols;
+using Nalix.Network.Routing;
 
-namespace Nalix.Network.Routing.Results.Memory;
+namespace Nalix.Network.Internal.Results.Memory;
 
 /// <inheritdoc/>
 [EditorBrowsable(EditorBrowsableState.Never)]
