+++ src/Nalix.Network/Internal/Results/Task/ValueTaskReturnHandler.cs
@@ -4,21 +4,23 @@
 using System.ComponentModel;
 using System.Threading.Tasks;
 using Nalix.Common.Networking.Packets;
+using Nalix.Network.Routing;
 
-namespace Nalix.Network.Routing.Results.Task;
+namespace Nalix.Network.Internal.Results.Task;
 
 /// <inheritdoc/>
 [EditorBrowsable(EditorBrowsableState.Never)]
-internal sealed class ValueTaskVoidReturnHandler<TPacket> : IReturnHandler<TPacket> where TPacket : IPacket
+internal sealed class ValueTaskReturnHandler<TPacket, TResult>(IReturnHandler<TPacket> innerHandler) : IReturnHandler<TPacket> where TPacket : IPacket
 {
     /// <inheritdoc/>
     public async ValueTask HandleAsync(object? result, PacketContext<TPacket> context)
     {
-        if (result is not ValueTask valueTask)
+        if (result is not ValueTask<TResult> valueTask)
         {
             return;
         }
 
-        await valueTask.ConfigureAwait(false);
+        TResult taskResult = await valueTask.ConfigureAwait(false);
+        await innerHandler.HandleAsync(taskResult, context).ConfigureAwait(false);
     }
 }
