+++ src/Nalix.Network/Internal/Results/Void/VoidReturnHandler.cs
@@ -4,8 +4,9 @@
 using System.ComponentModel;
 using System.Threading.Tasks;
 using Nalix.Common.Networking.Packets;
+using Nalix.Network.Routing;
 
-namespace Nalix.Network.Routing.Results.Void;
+namespace Nalix.Network.Internal.Results.Void;
 
 /// <inheritdoc/>
 [EditorBrowsable(EditorBrowsableState.Never)]
