+++ src/Nalix.Network/Internal/Routing/DispatchChannel.cs
