+++ src/Nalix.Network/Internal/Routing/PacketHandler.cs
@@ -8,9 +8,11 @@ using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 using System.Threading.Tasks;
 using Nalix.Common.Networking.Packets;
-using Nalix.Network.Routing.Results;
+using Nalix.Network.Internal.Results;
+using Nalix.Network.Routing;
+using Nalix.Network.Routing.Metadata;
 
-namespace Nalix.Network.Routing.Metadata;
+namespace Nalix.Network.Internal.Routing;
 
 /// <summary>
 /// Enhanced version of <c>PacketHandler</c> using compiled delegates for zero-allocation execution.
