+++ src/Nalix.Network/Routing/PacketDispatchOptions.PublicMethods.cs
@@ -14,9 +14,9 @@ using Nalix.Common.Exceptions;
 using Nalix.Common.Networking;
 using Nalix.Common.Networking.Packets;
 using Nalix.Network.Internal.Compilation;
+using Nalix.Network.Internal.Results;
+using Nalix.Network.Internal.Routing;
 using Nalix.Network.Middleware;
-using Nalix.Network.Routing.Metadata;
-using Nalix.Network.Routing.Results;
 
 namespace Nalix.Network.Routing;
