+++ src/Nalix.Network/Routing/PacketMetadata.cs
